install jsonwebtoken, route for connect users and see all the users in database
