+++ package.json
@@ -12,6 +12,7 @@
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "init": "^0.1.2",
+    "jsonwebtoken": "^8.5.1",
     "pg": "^8.2.1",
     "sequelize": "^5.21.13",
     "sequelize-cli": "^5.5.1"
