+++ routers/auth.js
@@ -1,42 +1,39 @@
-const { Router } = require("express")
-const bcrypt= require ('bcrypt')
+const {Router} = require ("express");
+const User= require ("../models").user;
+const bcrypt = require ('bcrypt')
 
-const User = require ("../models").user;
+const router = new Router()
 
+router.post("/", async ( req, res, next)=>{
+    // getparams => email and password validate
+    const {email,password} = req.body;
+    if(!email || !password){
+        res.status(400).send('missing login parameters')
+    }else{
+        const user = await User.findOne({where : {email}})
+        //look in db for user with the email
 
-const router = new Router();
+        if(!user){
+            res.status(404).send('User with that email adress not found')
+        }else {
+        // check if password match
 
-router.get("/", async (req,res,next)=>{
-    try{
-        const users = await User.findAll();
-        res.send(users)
-    }catch(error){
-        next(error)
-    }
-})
-
-router.post('/',async (req,res,next) => {
-    try{
-        const {email,password,age,name}=req.body
-
-        if(!password || !email || !age || !name){
-            res.status(400).send("missing some user parameters")
+        const passwordsMatch = bcrypt.compareSync(password, user.password)
+        if (passwordsMatch){
+            //guy exist have to log him in
+            console.log('right password');
+            
+            //create JWT
         }else{
-            const hashedPassword = bcrypt.hashSync(password, 10)
-
-            const newUser= await User.create({
-                email,
-                password : hashedPassword,
-                age,
-                name,
-            })
+            res.status(400).send("Wrong Password")
         }
- 
-        console.log(email,password,age,name);
-        res.send("creating a user")
-    }catch(error) {
-        next(error)
+        }
+
+        console.log('found user',user);
+        res.send("trying out login")
+       
     }
+    
 })
 
-module.exports = router;
+module.exports= router
